+++ packages/plugin-collapsible-list/web/src/components/modal-components/CollapsibleListSettings.jsx
@@ -4,6 +4,7 @@ import { mergeStyles } from 'wix-rich-content-common';
 import {
   RadioGroupVertical,
   SelectionList,
+  SelectionListItem,
   Separator,
   LabeledToggle,
   InfoIcon,
@@ -19,11 +20,8 @@ class CollapsibleListSettings extends Component {
     this.styles = mergeStyles({ styles, theme });
   }
 
-  renderOption = ({ item }) => (
-    <>
-      <item.icon />
-      <p>{item.label}</p>
-    </>
+  renderOption = ({ item, selected }) => (
+    <SelectionListItem icon={<item.icon />} selected={selected} label={item.label} />
   );
 
   renderExpandOptions = () => {
@@ -113,7 +111,6 @@ class CollapsibleListSettings extends Component {
           value={getDataManager().getDirection()}
           onChange={getDataManager().changeDirection}
           className={this.styles.direction_selector}
-          optionClassName={this.styles.direction_selector_option}
         />
       </>
     );